Moved base project settings into InheritanceProject config.
This was done to make sure that we have full control over how
getAssignedLabel() is called when configuring.

That in turn was done to change the behaviour of getAssignedLabel() to
always return a fully inherited and cached value.

That in turn was needed to speed up Queue.maintain(), which called the
label generation far too often for Reflection being usable.