Shared workspaces do not work, if the job has no "parameterized workspace" set
Shared workspaces currently depend on the "parameterized workspace" feature,
because of the call to: InheritanceBuild.getWorkspaceFor()

This function only looks at the parameterized workspace option, which
is wrong. It should check both "customized workspace" and the regular
workspace allocation.

Also, added a unittest for this use-case (which needed additional adjustments
to some other classes, like Reflection and InheritanceBuild).