Caching of labels leads to delayed reaction to label changes
The caching of label assignments to jobs has great benefits in Jenkins. It
greatly speeds up scheduling of jobs and prevents dangerous performance
degradations due to the strongly sequential nature of the Jenkins Queue
access.

The downside is, that changing labels on hosts is reacted to with a
significant delay.

Basically, labels only refresh the list of hosts they are valid for, when
the label is created for the first time, as it is an expensive operation.
This means, that cached labels fail to recognize label changes on hosts, or
when new hosts have been added into the system.

The previous workaround was to save an arbitrary project (as long as it is
an inheritance project type), which cleared the label cache.

The new approach should regenerate cached labels, by using Jenkins' internal
label cache lookup. This will preserve the speed savings of avoid expensive
inheritance, while using Jenkins' own caching for the actual resolution of
the contents of labels.