Synching source code between internal and open source state to 19.08.01
This commit rolls in all the Intel internal changes between March 2017 and
August 2019 into one commit.

This is to align the binary release with our internal code, before we switch
to a different repository, that will release this and other plugins in a
much more coherent and faster way.

The current way of concentrated code-dumps has simply not worked out as
desired. A much closer operation between the openly visible GitHub and
internal code repos needs to be reached and will be published soon.

Signed-off-by: Martin Schroeder <martin.schroeder@intel.com>