Fixed issue with CreationItemListener holding up the start of Jenkins
This is due to a race-condition of trying to create the ProjectCreationEngine
while the ExtensionLock is held by another Jenkins start thread.

That issue only affects some servers, due to the time-sensitive nature of
the locking mechanism. But if it happens, it happens reliably.

The solution is to delay the initialization of the PCE from *before* all
jobs are loaded, to *after* all jobs are loaded.

Also, this commit adds the ".groovy" path to the .gitignore. This path is
created by a the Groovy-Eclipse plugin.